refactor: extract file upload logic to separate module
